* fix a bug with the protocol version

+++ src/RDMHandler.cpp
@@ -216,8 +216,9 @@ void ResponseHandler::DeviceInfo(
     return;
 
   cout << "Device Info" << endl;
-  cout << "RDM Protocol Version: " << device_info.protocol_version_high <<
-    "." << device_info.protocol_version_low << endl;
+  cout << "RDM Protocol Version: " <<
+    static_cast<int>(device_info.protocol_version_high) <<
+    "." << static_cast<int>(device_info.protocol_version_low) << endl;
   cout << "Device Model: 0x" << std::hex << device_info.device_model << endl;
   cout << "Product Category: " <<
     ola::rdm::ProductCategoryToString(device_info.product_category) << endl;
